+++ src/main/java/com/mensal/sliceCtrl/service/IngredienteService.java
@@ -3,6 +3,7 @@ package com.mensal.sliceCtrl.service;
 import com.mensal.sliceCtrl.DTO.IngredientesDTO;
 import com.mensal.sliceCtrl.entity.Ingredientes;
 import com.mensal.sliceCtrl.repository.IngredienteRepository;
+import jakarta.transaction.Transactional;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -46,13 +47,26 @@ public class IngredienteService {
         return toIngredientesDTO(ingredientes);
     }
 
-
+    @Transactional
     public Ingredientes cadastrar(IngredientesDTO ingredientesDTO) {
+        Ingredientes ingredientesBanco = this.ingredienteRepository.findByNome(ingredientesDTO.getNomeIngrediente());
+        Assert.isTrue(ingredientesBanco == null, "Já possui um ingrediente com esse nome!");
+
         Ingredientes ingredientes = toIngredientes(ingredientesDTO);
         return this.ingredienteRepository.save(ingredientes);
     }
 
+    @Transactional
     public Ingredientes editar(IngredientesDTO ingredientesDTO, Long id) {
+
+        Ingredientes ingredientesBanco = this.ingredienteRepository.findByNome(ingredientesDTO.getNomeIngrediente());
+
+        if(ingredientesBanco != null){
+
+        if(ingredientesDTO.getId() != ingredientesBanco.getId()) {
+            Assert.isTrue(ingredientesBanco == null, "Já possui um ingrediente com esse nome!");
+        }
+        }
 
         final Ingredientes ingredientesBD = this.ingredienteRepository.findById(id).orElse(null);
         Assert.notNull(ingredientesBD, "Endereco inexistente!");
@@ -63,10 +77,21 @@ public class IngredienteService {
         return this.ingredienteRepository.save(ingredientes);
     }
 
+    @Transactional
     public void delete(final Long id) {
         final Ingredientes ingredientes = this.ingredienteRepository.findById(id).orElse(null);
-        Assert.notNull(ingredientes, "Endereco inexistente!");
-        this.ingredienteRepository.delete(ingredientes);
+        Assert.notNull(ingredientes, "Ingrediente inexiste!");
+
+
+        if (ingredientes != null) {
+            if (!ingredientes.getSabores().isEmpty()) {
+                throw new IllegalArgumentException("Não é possível excluir o ingrediente devido às relações com sabores existentes.");
+            } else {
+                this.ingredienteRepository.delete(ingredientes);
+            }
+            }
+
     }
 
 }
+
